requirements.txt
Appium-Python-Client==5.1.1
pytest==7.1.2
pytest-html==3.1.1
selenium==4.26.0